backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
cloudinary
python-multipart
pydantic-settings
pyjwt
httpx
qrcode
pillow
